Dont generate types for codemods, as its not a lib to import
