+++ packages/codemods/package.json
@@ -23,8 +23,8 @@
   "scripts": {
     "build": "yarn build:js",
     "prepublishOnly": "yarn build",
-    "build:js": "babel src -d dist --extensions \".js,.ts\" --copy-files --no-copy-ignored --ignore \"src/**/__tests__/**\" --ignore \"src/**/__testfixtures__/**\"",
-    "build:watch": "nodemon --watch src --ignore dist,template --exec \"yarn build\"",
+    "build:js": "babel src -d dist --extensions \".js,.ts\" --ignore \"src/**/__tests__/**\" --ignore \"src/**/__testfixtures__/**\"",
+    "build:watch": "nodemon --watch src --ignore dist --exec \"yarn build\"",
     "test": "jest",
     "test:watch": "yarn test --watch"
   },
