+++ packages/codemods/tsconfig.json
@@ -2,12 +2,10 @@
   "extends": "../../tsconfig.compilerOption.json",
   "compilerOptions": {
     "baseUrl": ".",
-    "rootDirs": ["src", "testUtils"],
-    "outDir": "dist",
+    "rootDir": "src",
+    "emitDeclarationOnly": false,
+    "noEmit": true
   },
-  "include": ["src", "./testUtils.d.ts", "testUtils"],
-  "exclude": ["**/__testfixtures__"],
-  "references": [
-    { "internal": "../internal" },
-  ]
+  "include": ["src", "./testUtils.d.ts"],
+  "exclude": ["**/__testfixtures__"]
 }
